backend/requirements.txt
# FastAPI + server
fastapi[standard]>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pgvector>=0.3
geoalchemy2>=0.15

# Cache + task queue
redis[hiredis]>=5.0
celery[redis]>=5.4

# Auth
pydantic-settings>=2.0
python-jose[cryptography]>=3.3

# HTTP client
httpx>=0.27

# ── Core module dependencies (shared with original app) ──
pandas>=2.1,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.4,<2.0
joblib>=1.3,<2.0
matplotlib>=3.8,<4.0
rapidfuzz>=3.6,<4.0

# Embeddings & LLM
sentence-transformers>=3.0,<4.0
transformers>=4.41,<5.0
huggingface-hub>=0.34,<1.0

# Ingestion
pypdf>=4.2,<5.0
faiss-cpu>=1.8,<2.0
python-docx>=1.1,<2.0
Pillow>=10.0,<12.0

# Spatial
shapely>=2.0,<3.0

# ML
xgboost>=2.0,<3.0
